move product img to assets folder, and fix imports
